+++ CHANGELOG.md
@@ -5,10 +5,9 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
-### Added
+### Internal
 
-- [#7](https://github.com/sunsided/async-tempfile-rs/pull/7):
-  Added `TempDir` for automatically deleted temporary directories.
+- Refactored temporary file name generation.
 
 ### Changed